refactor: OpenAPI schema 更新を関数へ切り出し
